requirements.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
chromadb
multion
llama-index
llama-index-vector-stores-chroma
macos_speech

git+https://github.com/openai/whisper.git
